add generic input type
