+++ src/handbrake.rs
@@ -3,4 +3,3 @@ mod model;
 
 pub use encoder::encode as encoder_with_handbrake;
 pub use model::HandbrakeInfo as HandbrakeInfo;
-pub use model::HandbrakeInputFile as HandbrakeInputFile;
